+++ js/data.js
@@ -1,0 +1,40 @@
+const getData = function(){
+  const nameMock = [
+    'Ivan',
+    'Fedor',
+    'Jack',
+    'Mark',
+    'ken',
+    'Bob',
+    'Alen',
+    'Den'
+  ];
+
+  const descriptionMock = [
+    'hello world',
+    'world is beautifull',
+    'life is good',
+    'sky is blue',
+    'sky is black',
+    'life is worst',
+    'goodbye world',
+    'where i am'
+
+  ];
+  const messageMock = [
+    'i am programmer',
+    'i am manager',
+    'i am shoes seller',
+    'i am engeneer',
+    'i am ofice worcker',
+    'i am driver',
+    'i am drug seller',
+    'i am doctor'
+
+  ];
+  return {nameMock,
+    descriptionMock,
+    messageMock
+  };
+};
+export{getData};
